do not leave unifinished processes on failure
This one allows all wget/curl processes to finish when catchup failure
is in progress. If any wget/curl process is left to finish, then it is
possible it will finish during TmpDir cleanup which could prevent it
from finishing. It may result in throwing an exception and crashing
whole stellar-core.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>