do not retry main catchup works
Retrying RECENT catchup work can cause problems if MINIMAL subwork
finished successfully and COMPLETE subwork did only part of its job. It
is better to let LedgerManager restart catchup completly than to let
this error happen. If COMPLETE did part of its work, then last closed
ledger is bigger than it was when MINIAMAL work started, so if MINIMAL
is executed next time with the same settings (as it does when RECENT is
restarted), an assert in ApplyBucketsWork will be issues and
stellar-core will crash.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>