do not pass references to ApplyBucketsWork
Passing reference to HistoryArchiveState& applyState may suggest that
this value will be changed inside ApplyBucketsWork and that this change
will be important to caller of ApplyBucketsWork constructor. That is not
true.

Also with non-obvious lifetime of Work classes it is dangerous to pass
references to them, as Work objects may outlife the references.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>