always catchup to first possible ledger instead of checkpoint
Now we can do catchup to any ledger we want, so it will be better to
catchup to ledger before one we got from network. It ensures that we
will always have ledger for verification in mSyncingLedgers in
LedgerManagerImpl so verification logic will be simplified and failures
caused by missing consensus in that buckets will no longer be possible.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>