only verify ledgers that are required to catchup
This is second part of ensuring that catchup failures will be even more
rare now. As we are now catching up not at checkpoint boundary but to
ledger previous to one that we got from network, it is no longer
required to verify ledgers in whole checkpoint. We can verify only those
ledgers that will be applied later, as last one of them can be verified
agains first one from mSyncingLedgers in LedgerManagerImpl.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>