simplify LedgerManagerImpl::verifyCatchupCandidate
Last part of ensuring that catchup failures are now very rare. In case
of manual catchup we dont have any buffered ledgers to verify catchup
candidate agains, so we don't do that.

In case of non-manual catchup logic of LedgerManagerImpl ensures that
catchup candidate can be verified agains first ledger in
mSyncingLedgers,
so we are just doing that.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>