publish queued history just after node is synced
Do not start publish work as soon as history entry has been added to
publishqueue table - do that after node is synced. Publish wont work
when node is not synced, it will just backoff PublishWork. It can cause
publish to be started more than hour later that it could be.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>