fix typo in admin.m
The word "to" is missing from the sentence. I had included a suggested replacement for the whole paragraph below.

It would be better to change the two sentences from:
Here, each player can decide to look wherever they want. Even look at only one
other player, which is a risk if this other player decides to quit the game.

To:
Here each player can decide to look wherever they want, even if they can only see one other player (or nobody). This is risky if that player quits the game or is not following the rules.