bucket: sink resolveAnyReadyFutures from HAS at ledger close to addBatch.
Formerly we resolved ready FutureBuckets on a HAS snapshot before storing
it into the database. This failed to resolve the FutureBuckets in the
BucketList itself and, worse, meant that earlier snapshot HASs like the
one taken by the history publishing system would not get resolved. This
change means that we resolve FutureBuckets on a regular basis "at their
source", such that any HASs that take snapshots will be (nearly) up to
date with the best-case state of resolution they can be.