reappend overriden -std=c++14 for enable-afl in configure.ac
This may seem like an arbitrary or counterintuitive change. Its root is
a change in the autotools ax_cxx_compile_stdcxx macro. This macro used
to append -std=c++XX to CXXFLAGS, but in 2015 this changed, it is now
appended to CXX. Stellar did not update the relevant lib file until
2018, and ever since then the fuzzer build has failed. This fix
replicates the actions of ax_cxx_compile_stdcxx. For a more detailed
explanation see: https://github.com/stellar/docker-stellar-core/pull/66#issuecomment-521886881