improve parallel build instructions
Add an explicit example of how to use `make -j` to run a parallel
builds, i.e. `make -j<N>`, or for example, `make -j3`.

The install instructions suggest using `make -j` for parallel builds but
if this command is executed without a limit being provided to the `j`
option, it may lock up the machine it is being run on. For those not
familiar with the `-j` option it is a bit of a trap. Changing the
example command to show the use of the limit should prevent folks new to
building this repo and unfamiliar with the `-j` option from making that
mistake.