{Ballot,Nomination}Protocol::processEnvelope: Use early return to simplify flow
This diff is better viewed with whitespace changes ignored.

By using early returns, we simplify the code flow,
making it more obvious which path are error path and which are not.
It also lessen the burden on the programmer to keep context in mind,
and reduce the indenting level as an added bonus.

While NominationProtocol was fairly consistent in its usage of a 'res' value,
BallotProtocol used a mixed of both styles.
Not that some error handling has also been hoisted as a result of the refactoring.