Initial support for DB extension-schema changes for #2570
Database schema change to try to allow future additions to
extension fields of AccountEntry and TrustLineEntry, including
some impending for CAP-33, without further database schema changes:

- Bump the DB schema version number.

- When upgrading to the new DB schema, transfer the contents of
any non-NULL Liabilities fields in AccountEntry and TrustLineEntry
tables into a new "extension" column, which is opaque to the DB
code.  Then remove the old "buyingliabilities" and
"sellingliabilities" columns (if using PostgreSQL).

- Replace explicit uses of Liabilities structure (which consists
of two INT64s, "buyingliabilities" and "sellingliabilities")
throughout AccountEntry and TrustLineEntry database operations
with uses of an opaque "extension v1" structure.  By making
the next extension, v2, as proposed in CAP-33, into an extension
of extension v1, and v3 as an extension of v2, and so on, this
should allow future additions to the XDR of the extension fields
without changes to the DB schema or SQL code.

- Add (not-yet-used) test utility functions which create the
"accounts" and "trustlines" databases as they are now, in case
we want to continue using them unchanged in unit tests in the
future even if we remove support in the production code for
the current schema.  (Note that this checkin is not actually
changing the production table creation code, because that code
path starts by creating the database with an earlier schema
version and upgrading the newly-created old-schema database
step-by-step to the current schema.)