DB schema upgrade to make extensions opaque
Upgrade the DB schema from 12 to 13.  The new schema makes
all of the extensions in all ledger entry types opaque to
the database code itself.  Those extensions are the
extension to LedgerEntry itself, which is stored in the
table for each LedgerEntry type (account, trustline, data,
offer), and the extensions which are specific to each type.

The idea of making the extensions opaque is that they can
be extended in the future without further changes to either
the DB schema or the DB code, as the database itself will
simply continue to treat the extensions as variable-length,
opaque XDR.  The cost is that we can't query on components
of the extensions (which we don't do currently, and won't
do with any of the extensions that we currently have
planned).

The general LedgerEntry extension, as well as the extensions
to DataEntry and OfferEntry, are currently unused -- they
are unions with the only version being 0 and no content
besides that fixed version number, and they are not stored
in any database columns at all.  The upgrade code therefore
simply creates columns for them and populates all of the
cells of the new columns with base-64 encodings of opaque
XDR representing that trivial tagged union.

The two extensions that are already used are those specific
to AccountEntry and TrustLineEntry, both of which contain
the same thing, namely a Liabilities structure (which
contains two fields, one for buying and one for selling).
Those extensions are tagged unions with a genuine version
option -- version 0 for entries that have never yet had
liabilities, and then version 1 when liabilities are first
created for the entry.  Existing databases have NULL in
the "buyingliabilities" and "sellingliabilities" columns
for the version-0 entries, and non-NULL values, namely
base-64 encodings of the XDR for the Liabilities structure,
for the version-1 entries.

Therefore, the upgrade code for AccountEntry and
TrustLineEntry performs a conversion:  for each record,
it reads the "buyingliabilities" and "sellingliabilities"
columns, which are 64-bit integers transparent to the
database, and generates encoded XDR, opaque to the database,
representing an extension field which, in those cases,
has a Liabilities field, and has the values read from the
old columns filled in.  (If the old liabilities fields for
a record are NULL, then the new "extension" column for that
record is also NULL -- the new live database code will
recognize that NULL values of those two extensions are
possible.)

Once all of the records have had their new "extension"
columns written, the upgrade code deletes the old
"buyingliabilities" and "sellingliabilities" columns.