Test upgrades to database schema 13
Test upgrades to database schema 13, in which all
extensions in all ledger entry types become opaque XDR
as far as the database is concerned.

The tests use the test-code-injection mechanism which allow
them to run arbitrary code between the creation of a
database (which is done in an old schema version,
MIN_SCHEMA_VERSION) and the upgrading of that database to
the current schema version (SCHEMA_VERSION).  They create
ledger entries of each of the four types, using raw SQL so
that they can create them in the old format which the live
production code no longer uses.  After the database upgrade
completes (which it has done by the time that
createTestApplication() returns), the tests validate that
the contents of the ledger entries are intact and
logically unchanged (although their representations in
the database have changed, with some columns having been
created and some having been deleted, and some contents
being encoded differently).