Upgrade protocol for CAP-0034 / protocol issue 622
This commit implements the protocol change documented in CAP-0034:

https://github.com/stellar/stellar-protocol/blob/master/core/cap-0034.md

(It does not bump the protocol version, because the version
has already been bumped since the last stellar-core
release.)

That proposal came from the discussion of stellar-protocol
issue #622:

https://github.com/stellar/stellar-protocol/issues/622

The CAP enumerates all the semantic changes in this commit.
Here are a couple of implementation notes too specific to
the code to be mentioned in the CAP:

- The interfaces that had an "upperBoundCloseTimeOffset"
added by PR 2608 have now also been extended with
"lowerBoundCloseTimeOffset"s, which analogously affect
the conditions under which isTooEarly() returns true,
allowing the new CAP-34 code to validate minTime as well
as maxTime (the latter uses "upperBoundCloseTimeOffset")
against the next close time (rather than the last ledger
close time).

- testSCPDriver() has been not only extended to test new
behaviors, but also refactored to allow more test cases to
be added by adding more test case parameters to a vector.