Remove obsolete protocol-10 test code
Remove obsolete "withSCPsignature" parameter from
testSCPDriver(); it is always true following the removal
of support for protocols which did not use signed values
during the nomination phase.

Remove obsolete protocol-10 call to testSCPDriver().

Remove "checkHighFee" parameter to testSCPDriver() which
became obsolete as a result of the removal of the
protocol-10 call.