herder: don't reject messages as aggressively during bootstrap
In test cases we have situations where we bootstrap a new network,
and lose sync within the first few ledgers (before MAX_SLOTS_TO_REMEMBER=12 ledgers).
When this happens, we want to ensure that we continue to process SCP messages,
regardless of their timestamp.

The old code was approximating "never been in sync" by just looking at the ledger
window, which is inacurate.