Implement enhanced manual close
PR 2696:
--------

Allow some configuration parameters that currently can't be
set along with NODE_IS_VALIDATOR if RUN_STANDALONE is also
set.

Return a string from Application::manualClose(), to allow
returning more specific messages (both upon success and upon
failure) to clients.

Add optional parameters to manualClose() for the new
"enhanced" parameters to the /manualclose endpoint.  Allow
those parameters to be specified only if RUN_STANDALONE is
configured (in addition to MANUAL_CLOSE).  Impose a limit
on the latest close time to the beginning of 2200, which
suffices to prevent overflows in Herder.

Implement enhanced manual close in
ApplicationImpl::manualClose() by making the requested
manual modifications around the call to
HerderImpl::triggerNextLedger().

PR 1760:
--------

Use VIRTUAL_TIME when MANUAL_CLOSE and RUN_STANDALONE are
both configured.  When they are, ensure that automatic
maintenance is disabled so that it doesn't continually
increase virtual time in a tight loop of trimming history;
also ensure that Herder does not spin in a tight loop
rebroadcasting (the latter is disabled any time that
RUN_STANDALONE is configured, regardless of MANUAL_CLOSE).

When MANUAL_CLOSE and RUN_STANDALONE are configured, and a
manual close is performed without explicitly specifying a
new closeTime, default to the real system clock time, even
though stellar-core is running in VIRTUAL_TIME, to preserve
the current behavior for existing "manualclose" clients.