Replace maxCommit() = (N / 2) + 1 with N / 2 for possibly faster quorum intersection checks
This should still work because:
* Let N be the number of nodes in the network.
* Let A, B be disjoint quorums.

Let k be an integer such that N = 2k or 2k + 1 depending on the parity of N.
If |A| > k and |B| > k, then |A| >= k + 1 and |B| >= k + 1.
Therefore, |A| + |B| >= 2k + 2 > N.
This is impossible, so at least one of A or B has at most k elements.
In other words, floor(N / 2) elements.