Use root account consistently in fuzzer setup
Fuzzer setup has been special-casing pregenerated account 0
in various places, and using it as the source account for
transactions.  In this commit we switch to using the root
account as the source account, as well as the target for
payments that reduce the balances of other accounts after
fuzzer setup, so that we can stop special-casing account 0.