Map ClaimableBalanceID and LedgerKey to one byte
Compact `ClaimableBalanceID`s and `LedgerKey`s into a
single byte, and use that byte to index into an array of
keys generated after setup, in the hope of reducing the
space that the fuzzer has to explore in order to perform
operations on sponsorships and claimable balances created
during setup.  In particular, it shouldn't have to guess a
hash anymore.

This comprises the following changes:

- After setup completes, store the ledger entries that it
  created in an array in the TransactionFuzzer, which the
  XDR unpacker can use single-byte ClaimableBalanceIDs and
  LedgerKeys to index into.  Fill the leftover space in the
  array with pseudorandomly generated entries (some
  validated, some not).
- In the XDR compactor, convert `LedgerKey` and
  `ClaimableBalanceID` into a single byte (the first
  byte of the opaque XDR representation).
- In the XDR unpacker, unpack `LedgerKey`s and
  `ClaimableBalanceID`s into ones from the stored array of
  `LedgerKey`s.