Order-book-not-crossed invariant implementation
A new invariant which keeps track of the state of the order
book as operations are applied, and asserts that it is not
crossed.

There is currently no mechanism by which an invariant can be
informed that an operation application for which its
checkOnOperationApply() method had been called is being
rolled back.  Hence, we can not in general support an
invariant such as this which maintains state across method
calls.  But we will be able to do so when fuzzing, where
setup operations are never rolled back and where we can
explicitly tell invariants when to snapshot state (that is,
after setup completes) and when to reset it to that state
(after each fuzz test).