Wrap test uses of TransactionFrameBase::checkValid
Introduce a txtest::checkValid() which tests can use as a
wrapper around calls to TransactionFrameBase::checkValid().
Currently all it provides are some default parameters (most
tests pass all 0's), but it will be used later to validate
the relationship between the return values of different
(yet-to-be-defined) forms of
TransactionFrameBase::checkValid().