Use expected hash and meta config in ledger close
Detect corruption during ledger close when the client
provides an expected hash in LedgerCloseData.

At the beginning of ledger close, if there is still meta
to emit from the previous ledger, emit it.

At the end of ledger close, if either UNSAFE_EMIT_META_EARLY
is set or the client provided an expected ledger hash which
the actual hash matches, emit the metadata from the ledger
close.