SCPDriver: Make to{ShortString,StrKey} accept NodeID
As SCP is intended to be a library, the crypto dependency was mostly removed from SCP
in PR2048, but some parts lingered. In particular, those two methods, used only for
debugging, bring in a lot of dependencies.
In order to reduce those dependencies, the first step is to change them to refer
to the type used to identify a node (NodeID), as opposed to its implementation.