Rename isString32Valid to isStringValid
### What
Rename isString32Valid to isStringValid.

### Why
isString32Valid is named with 32 in the name which implies that it is
validating 32 byte strings, but it is used to validate strings of both 32 bytes
and 64 bytes in length. For example, it is used to validate 32 byte home
domains, and 64 byte data entry names. When perusing the code this initially
confused me and I thought the data entry names must also be 32 bytes. On closer
inspection I discovered that the isString32Valid does not check if the string
is valid for the string32 type as specified in the Stellar XDR, but rather that
the string is valid for specific string fields in the XDR, namely the home
domain and the data entry name.

It would be less misleading if this function didn't specify the string size in
the name when it is not relevant to the functions behavior or intended scope of
use.