Use $(Configuration) variable for the paths of generated files (XDR and Rust).
While this works correctly for builds, Visual Studio doesn't seem to be capable of updating the solution index when the configuration changes. So generated files browsing may look incorrectly after configuration changes. The best workaround I have found is to just restart VS after the config change (most of the time there is no need to browse generated files, hence this is not a big issue).