Replace `generateLoad` arguments with a single config struct.
This currently has a significant number of arguments, some of which are only modified in a single place. The following changes need even more parameters, so maintaining all the args/defaults becomes unwieldy.

This will look nicer with struct designators in C++20, currently I'm using factory functions to address the common patterns.