Introduce parameters to load generator to skip low fee txs and generate diverse fees.
While it's currently possible to trigger surge pricing with load generator, the request need to fit exactly into 1x and 2x the tx set size, which is tricky if we want to also randomize operation counts.

This mode also allows to exercise the tx queue limiter, which is useful too.