Make tx set frame order matter only during encoding/decoding.
There is no reason to maintain the order in memory and then implicitly assume it's preserved during encoding. Instead, we store transactions in arbitrary order in memory and sort only during encoding to XDR. This hopefully should reduce the risk of outputting a malformed tx set.

For additional safety, we do the XDR roundtrip during the in-memory tx set construction, so we guarantee that we never nominate malformed tx set XDR (at least from the standpoint of the particular core version).