Overlay: Do not initiate connections that will be immediately dropped
This change ensures that the overlay does not initiate outbound
connections when it is shutting down (and therefore
`OverlayManagerImpl::addOutboundConnection` will immediately drop the
connection).

This change also adds a line to the logs indicating how to increase the
number of outbound pending slots when
`OverlayManagerImpl::connectToImpl` fails due to insufficient slots.