Add standalone `check-quorum-intersection` command
Closes #4062

This change adds a command `check-quorum-intersection` that takes a JSON
file produced by the `quorum` endpoint and checks whether it enjoys a
quorum intersection.

Most of the change centers around reading JSON in, but this change also
required refactoring `QuorumIntersectionChecker` to use a map from
`NodeID` to `SCPQuorumSetPtr` instead of a map from `NodeID` to
`NodeInfo`, as the intersection checking only relies on the `mQuorumSet`
field of a `QuorumMap` and the JSON input does not contain everything in
`NodeInfo`.