Draft: Recover SCP messages when replaying from history
This PR enables updating the `scphistory` table during catchup from
history. It allows users to specify which archives to use via the
`SCP_HISTORY_ARCHIVES` config option. If a user specifies multiple
archives, stellar-core will merge the messages from the archives.

This is a draft PR because I'm looking for feedback on the approach, but
still have some work to do before it is in a mergeable state.
Most of the remaining work is to:

* Document the new functionality
* Write additional tests for:
  * Merging
  * Failed downloads
  * No `SCP_HISTORY_ARCHIVES`
  * Multiple `SCP_HISTORY_ARCHIVES`
  * Changes to `ReplayDebugMetaWork`
* Integrate changes from #4121
* Address other `TODO`s in the changes