Add `simulate` subcommand to network survey script
This change adds a `simulate` command to the network survey script that
takes a network topology in graphml format (such as the one produced by
the `survey` command) and simulates a survey of said network.  It is
intended as a tool to use during development of the script itself to
test the script without running a survey on the network, and should
make it easier to address issues such as #2592.

This is an initial version of the `simulate` command that will evolve as
we need to test the script under additional scenarios.