Reduce the range of the testnet write fees.
Due to small bucket list size, even relatively minor changes in testnet (+/- 20 MB) cause significant fee shifts and as a result fees that are significantly higher than on pubnet. We can't achieve perfect synchronization between the network, but at least with this change the range is reduce to [0.9, 2] times of the initial phase 2 write fee, which hopefully will be relatively close to the pubnet fee for a while.