Update network survey script to support V2 survey
Resolves #4302

This change updates the network overlay survey script to support the new
V2 survey API.

To test this, I also updated the simulator to simulate the new API, then
I checked that the resulting graph from simulating an old pubnet
survey is structurally isomorphic to the input graph. I also checked
that the new fields are present in the output graph.