Fix MSVS project.
It seems like it's not possible to link the release build of the rust libraries to the debug-mode core, so I'm disabling tracy for the debug builds. Neither debug rust builds, nor tracy in debug mode are useful on Windows builds (FWIW I'm not sure if tracy is usable at all on Windows, but in general profiling shouldn't be normally done on debug builds).

There is also a weird issue of `intrin.h` missing when tracy is disabled, so I've added the include explicitly.