Update soroban p21 version to 21.2.2, taking binary-search change.
This is a critical update. Stellar-core packages from before this
update, if built with rust 1.82, will diverge from the network. They
must not be deployed in production validators.

Rust 1.82 changed the implementation of binary search in its standard
library in a way that soroban observes. This change modifies soroban to
use a local copy of binary search that is the same as the one in rust
1.81, so that soroban is no longer sensitive to the difference between
1.81 and 1.82.

Existing SDF-built release packages of stellar-core were built with rust
1.81, but it is possible that other builds of stellar-core built after
the release of rust 1.82 might exist in the field. These must all update
to this change, or later, to avoid diverging from peers built with rust
1.81.