Error and exit on validator configs that contain only `LOW` validators
Closes #4526

This change raises an error when a node is a validator and all of the
validators in its config (including itself) are marked `LOW` quality.
This represents a misconfiguration because `LOW` quality validators
cannot win leader election, and so if every validator is `LOW` quality,
then there are no eligible nomination leader.