Add script to generate transaction histograms from Hubble data
This change adds a script that takes Hubble transaction data as input
and outputs distributions over transaction resource usage to simplify
the updating of loadgen distributions in supercluster.

This is the same script originally included in
stellar-core/supercluster#220. As per
[this comment](https://github.com/stellar/supercluster/pull/220#discussion_r1849395640),
I've replaced instances of `tx_history` with the more accurate
`history_transactions`, which matches the name of the underlying Hubble
table.