Fix tests in `--disable-postgres` build
Resolves #4625 and #4614.

This change fixes the tests in the `--disable-postgres` build by
disabling parallel ledger close tests when postgres is disabled. This is
necessary because parallel ledger close requires postgres.

As part of this, I also removed the
`TESTDB_BUCKET_DB_PERSISTENT_POSTGRES` test mode, as it's made redundant
by the `TESTDB_POSTGRESQL` mode.