Remove support for old V1 survey
Closes #4332.

Now that all stellar-core nodes on the network support the V2 survey,
and there are no node stellar-core nodes that will generate V1 survey
messages, it's safe to cleanup the codebase a bit and remove support for
the old V1 survey entirely. This has the benefit of greatly simplifying
`SurveyManager`.