Adapt Core to the CAP-67 XDR.
This is unfortunately rather hacky in order to allow for these changes to go in a minor release. Specifically, we're only bumping XDR for the Core itself, and not for the Rust env. That's why we need to temporarily disable a few XDR integrity checks. There is also a point of interaction of the new and old XDR(ScAddress changes), and I've added some tests to ensure that Core won't be able to accidentally accept XDR that's not supported by the current protocol.