More fixes to fee bump tx tests.
The apply tests were doing something really weird: they used the result generated by `checkValidForTesting` and then proceeded to `apply`. This is not a flow that actually exists in Core - `apply` always uses the result provided by `processFeeSeqNum` and `checkValid` is completely separate from the apply flow. This issue resulted in observing the inner fee charged as 0 and not the inner tx fee charged, i.e. it hid the fact that we always set `feeCharged` in the inner result of fee bump txs.

Also updated the whole test to not use `checkValidForTesting` and removed an unnecessary argument from the event manager interface.