Increase refundable fee in Soroban tests and regenerate meta.
This is done in order to get clean meta diffs for the rent fee calibration PR: it requires the fee increase, but is also a protocol change, which is why it's important to distinguish between the real meta diffs and test setup changes.

This also contains passing-by test improvements and flakiness fixes.