Track Soroban in-memory state size and store it in p23.
This adds the state size tracking mechanism to InMemorySorobanState. The size includes the size of the parsed Wasm modules and thus has to be recomputed on the protocol and relevant setting upgrades. We'll start tracking the state size as soon as we're in Soroban protocol, but will only use it in p23. Because of that we can use p23 host to compute the memory cost for contract code entries eagerly.

This does a few non-trivial things on upgrades:

- When we perform a memory cost settings upgrade (even in p22) we recompute the contract code in-memory size. If we are in p23+, we also fill the state size window with the current in-memory state size.
- When we upgrade to p23, we immediately fill the state size window with the current in-memory state size, but don't recompute the contract code size (it's already up to date)
- When we upgrade to p24+, we recompute the contract code size and fill the state window with the current in-memory state size

All these cases, as well as the size tracking itself, are covered in the respective new tests.