Refactor SorobanNetworkConfig to be immutable and reduce its caching.
Also separated config updates from the in-memory config.

The motivation of the changes is as follows:

- Config is semantically not mutable. It's just a view for the ledger snapshot that is convenient to pass around and sometimes store.
- Config caching during apply resulted in some bugs (like broken updates for some state archival settings). Loading the config is cheap enough, so we just load it from LTX during the apply process.
- Updates are performed on the ledger entries and they don't need to be coupled with the view data structure.

On-demand config loading is still not 100% fool-proof and it's still possible to accidentally access stale config during the apply flow. However, I don't have a better idea for how this could be further improved at the moment. But at least outside of the apply flow only the last closed config can be accessed now, which I think is a net win for the change.