Fix pretty printing for Catch2.
Since Catch2::StringMaker specializations have been defined in a separate header, the wrong include order of that header would lead to ODR violation, that caused some (most?) tests to fallback to the default StringMaker defined in catch.hpp.

The fix just makes it so all the tests just include Catch2 via the header that also defines the StringMakers. I believe this should prevent ODR violations, as long as we use the new header for including Catch2 consistently.