Fix flaky "apply generated parallel tx sets" test.
There were some setup issues (a bunch of BAD_SEQ errors due to reusing accounts), but in general with the way is set up it's not possible to always included all the txs into ledger given multiple stages and cluster. So I've relaxed the check to allow up to 2 txs to not be included.