Remove obsolete test (that happens to do AC_RUN and breaks some environments)
This test is obsolete in the first place (we don't build on clang 8 at all
anymore) but also, incidentally, since it runs AC_RUN_IFELSE it winds up
executing a binary it built, and this happens after asan is enabled, and so in
a few specific environments where asan doesn't work (eg. old versions of docker
when running under qemu-user; also seemingly pbuilder) this causes
configuration to fail entirely.

There's no real way to make asan work in these environments, we have to just
avoid running conftests after asan is turned on. We could move this test above
the asan enabling part of configure, but again, it's an obsolete test; simpler
just to remove it entirely.