Increase granularity of SAC max TPS test.
There is really no need to measure it up with the granularity of 1 transaction, especially given that the variance is quite high. Also there shouldn't be a need to use the ledger close time that's not a multiple of 1 second, given that the test outputs TPS and not the precise ledger limits.