Decrease the max entry TTL to ~4 weeks for testnet.
There is a lot of contract prototyping going on in testnet which results in rather fast in-memory state growth, which is unbounded as the fees are meaningless for testnet. With the max extension reduced it won't be possible to keep state on-chain without actively using and bumping it, which hopefully should get most of the outdated prototypes to expire rather quickly.

This might result in some contracts failing on testnet if a temporary entry is being extended for more than allowed by the network limit. However, it's not clear if this is going to actually affect anyone, as the temporary entries are either very short-lived, or have TTL that is configured by the respective call (like allowance or signature TTL), which should be easy to workaround in testing. If this proves to be an issue, we'll need to add a separate configuration setting for the temp entry maximum TTL.

This is to be used with the next testnet reset that happens very soon; it wouldn't affect testnet currently.